requirements.txt
coverage==4.3.1
Cython==0.25.2
Django==1.10.5
django-pytest==0.2.0
django-static-precompiler==1.5
djangorestframework==3.5.3
mistune==0.7.3
pluggy==0.4.0
psycopg2==2.6.2
py==1.4.32
Pygments==2.1.3
pytest==3.0.5
pytest-cov==2.4.0
tox==2.5.0
